ai-story-teller-backend-python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic>=2.9.0
pydantic-settings>=2.1.0
email-validator
motor
beanie
openai>=1.50.0
assemblyai
jiwer
nltk
boto3
chromadb>=0.5.0
langchain
langchain-openai
langchain-community
langchain-chroma
langsmith
python-dotenv
pytest
pytest-asyncio
httpx
mongomock_motor
PyPDF2
ebooklib
beautifulsoup4
langgraph
langsmith
